split dashboard file into multiple files
