+++ pkg/triggers/dashboard/dashboard.go
@@ -1,24 +1,15 @@
 package dashboard
 
 import (
-	"context"
 	"fmt"
-	"sort"
-	"strings"
 	"sync"
 	"time"
 
-	"github.com/Notifiarr/notifiarr/pkg/apps"
 	"github.com/Notifiarr/notifiarr/pkg/plex"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/common"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/plexcron"
 	"github.com/Notifiarr/notifiarr/pkg/website"
 	"golift.io/cnfg"
-	"golift.io/starr"
-	"golift.io/starr/lidarr"
-	"golift.io/starr/radarr"
-	"golift.io/starr/readarr"
-	"golift.io/starr/sonarr"
 )
 
 /* This file sends state of affairs to notifiarr.com */
@@ -249,696 +240,6 @@ func (c *Cmd) getStatesParallel() *States { //nolint:funlen
 	wg.Wait()
 
 	return states
-}
-
-func (c *Cmd) getDelugeStates() []*State {
-	states := []*State{}
-
-	for instance, app := range c.Apps.Deluge {
-		if app.Timeout.Duration < 0 || app.URL == "" {
-			continue
-		}
-
-		c.Debugf("Getting Deluge State: %d:%s", instance+1, app.URL)
-
-		state, err := c.getDelugeState(instance+1, app)
-		if err != nil {
-			state.Error = err.Error()
-			c.Errorf("Getting Deluge Data from %d:%s: %v", instance+1, app.URL, err)
-		}
-
-		states = append(states, state)
-	}
-
-	return states
-}
-
-func (c *Cmd) getLidarrStates() []*State {
-	states := []*State{}
-
-	for instance, app := range c.Apps.Lidarr {
-		if app.Timeout.Duration < 0 || app.URL == "" {
-			continue
-		}
-
-		c.Debugf("Getting Lidarr State: %d:%s", instance+1, app.URL)
-
-		state, err := c.getLidarrState(instance+1, app)
-		if err != nil {
-			state.Error = err.Error()
-			c.Errorf("Getting Lidarr Queue from %d:%s: %v", instance+1, app.URL, err)
-		}
-
-		states = append(states, state)
-	}
-
-	return states
-}
-
-func (c *Cmd) getRadarrStates() []*State {
-	states := []*State{}
-
-	for instance, app := range c.Apps.Radarr {
-		if app.Timeout.Duration < 0 || app.URL == "" {
-			continue
-		}
-
-		c.Debugf("Getting Radarr State: %d:%s", instance+1, app.URL)
-
-		state, err := c.getRadarrState(instance+1, app)
-		if err != nil {
-			state.Error = err.Error()
-			c.Errorf("Getting Radarr Queue from %d:%s: %v", instance+1, app.URL, err)
-		}
-
-		states = append(states, state)
-	}
-
-	return states
-}
-
-func (c *Cmd) getReadarrStates() []*State {
-	states := []*State{}
-
-	for instance, app := range c.Apps.Readarr {
-		if app.Timeout.Duration < 0 || app.URL == "" {
-			continue
-		}
-
-		c.Debugf("Getting Readarr State: %d:%s", instance+1, app.URL)
-
-		state, err := c.getReadarrState(instance+1, app)
-		if err != nil {
-			state.Error = err.Error()
-			c.Errorf("Getting Readarr Queue from %d:%s: %v", instance+1, app.URL, err)
-		}
-
-		states = append(states, state)
-	}
-
-	return states
-}
-
-func (c *Cmd) getQbitStates() []*State {
-	states := []*State{}
-
-	for instance, app := range c.Apps.Qbit {
-		if app.Timeout.Duration < 0 || app.URL == "" {
-			continue
-		}
-
-		c.Debugf("Getting Qbit State: %d:%s", instance+1, app.URL)
-
-		state, err := c.getQbitState(instance+1, app)
-		if err != nil {
-			state.Error = err.Error()
-			c.Errorf("Getting Qbit Data from %d:%s: %v", instance+1, app.URL, err)
-		}
-
-		states = append(states, state)
-	}
-
-	return states
-}
-
-func (c *Cmd) getSonarrStates() []*State {
-	states := []*State{}
-
-	for instance, app := range c.Apps.Sonarr {
-		if app.Timeout.Duration < 0 || app.URL == "" {
-			continue
-		}
-
-		c.Debugf("Getting Sonarr State: %d:%s", instance+1, app.URL)
-
-		state, err := c.getSonarrState(instance+1, app)
-		if err != nil {
-			state.Error = err.Error()
-			c.Errorf("Getting Sonarr Queue from %d:%s: %v", instance+1, app.URL, err)
-		}
-
-		states = append(states, state)
-	}
-
-	return states
-}
-
-func (c *Cmd) getDelugeState(instance int, app *apps.DelugeConfig) (*State, error) { //nolint:funlen,cyclop
-	start := time.Now()
-	xfers, err := app.GetXfersCompat()
-	state := &State{
-		Elapsed:  cnfg.Duration{Duration: time.Since(start)},
-		Instance: instance,
-		Name:     app.Name,
-		Next:     []*Sortable{},
-		Latest:   []*Sortable{},
-	}
-
-	if err != nil {
-		return state, fmt.Errorf("getting transfers from instance %d: %w", instance, err)
-	}
-
-	for _, xfer := range xfers {
-		if eta, _ := xfer.Eta.Int64(); eta != 0 && xfer.FinishedTime == 0 {
-			//			c.Error(xfer.FinishedTime, eta, xfer.Name)
-			state.Next = append(state.Next, &Sortable{
-				Name: xfer.Name,
-				Date: time.Now().Add(time.Second * time.Duration(eta)),
-			})
-		} else if xfer.FinishedTime > 0 {
-			seconds := time.Duration(xfer.FinishedTime) * time.Second
-			state.Latest = append(state.Latest, &Sortable{
-				Name: xfer.Name,
-				Date: time.Now().Add(-seconds).Round(time.Second),
-			})
-		}
-
-		state.Size += int64(xfer.TotalSize)
-		state.Uploaded += int64(xfer.TotalUploaded)
-		state.Downloaded += int64(xfer.AllTimeDownload)
-		state.Downloads++
-
-		if xfer.UploadPayloadRate > 0 {
-			state.Uploading++
-		}
-
-		if xfer.DownloadPayloadRate > 0 {
-			state.Downloading++
-		}
-
-		if !xfer.IsFinished {
-			state.Incomplete++
-		}
-
-		if xfer.IsSeed {
-			state.Seeding++
-		}
-
-		if xfer.Paused {
-			state.Paused++
-		}
-
-		if xfer.Message != "OK" {
-			state.Errors++
-		}
-	}
-
-	sort.Sort(dateSorter(state.Next))
-	sort.Sort(sort.Reverse(dateSorter(state.Latest)))
-	state.Next.Shrink(showNext)
-	state.Latest.Shrink(showLatest)
-
-	return state, nil
-}
-
-func (c *Cmd) getLidarrState(instance int, app *apps.LidarrConfig) (*State, error) {
-	state := &State{Instance: instance, Next: []*Sortable{}, Name: app.Name}
-	start := time.Now()
-
-	albums, err := app.GetAlbum("") // all albums
-	state.Elapsed.Duration = time.Since(start)
-
-	if err != nil {
-		return state, fmt.Errorf("getting albums from instance %d: %w", instance, err)
-	}
-
-	artistIDs := make(map[int64]struct{})
-
-	for _, album := range albums {
-		have := false
-		state.Albums++
-
-		if album.Statistics != nil {
-			artistIDs[album.ArtistID] = struct{}{}
-			state.Percent += album.Statistics.PercentOfTracks
-			state.Size += int64(album.Statistics.SizeOnDisk)
-			state.Tracks += int64(album.Statistics.TotalTrackCount)
-			state.Missing += int64(album.Statistics.TrackCount - album.Statistics.TrackFileCount)
-			have = album.Statistics.TrackCount-album.Statistics.TrackFileCount < 1
-			state.OnDisk += int64(album.Statistics.TrackFileCount)
-		}
-
-		if album.ReleaseDate.After(time.Now()) && album.Monitored && !have {
-			state.Next = append(state.Next, &Sortable{
-				id:   album.ID,
-				Name: album.Title,
-				Date: album.ReleaseDate,
-				Sub:  album.Artist.ArtistName,
-			})
-		}
-	}
-
-	if state.Tracks > 0 {
-		state.Percent /= float64(state.Tracks)
-	} else {
-		state.Percent = 100
-	}
-
-	state.Artists = len(artistIDs)
-	sort.Sort(dateSorter(state.Next))
-	state.Next.Shrink(showNext)
-
-	if state.Latest, err = c.getLidarrHistory(app); err != nil {
-		return state, fmt.Errorf("instance %d: %w", instance, err)
-	}
-
-	return state, nil
-}
-
-// getLidarrHistory is not done.
-func (c *Cmd) getLidarrHistory(app *apps.LidarrConfig) ([]*Sortable, error) {
-	history, err := app.GetHistoryPage(&starr.Req{
-		Page:     1,
-		PageSize: showLatest + 20, //nolint:gomnd // grab extra in case some are tracks and not albums.
-		SortDir:  starr.SortDescend,
-		SortKey:  "date",
-		Filter:   lidarr.FilterTrackFileImported,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("getting history: %w", err)
-	}
-
-	table := []*Sortable{}
-	albumIDs := make(map[int64]*struct{})
-
-	for _, rec := range history.Records {
-		if len(table) >= showLatest {
-			break
-		} else if albumIDs[rec.AlbumID] != nil {
-			continue // we already have this album
-		}
-
-		albumIDs[rec.AlbumID] = &struct{}{}
-
-		// An error here gets swallowed.
-		if album, err := app.GetAlbumByID(rec.AlbumID); err == nil {
-			table = append(table, &Sortable{
-				Name: album.Title,
-				Sub:  album.Artist.ArtistName,
-				Date: rec.Date,
-			})
-		}
-	}
-
-	return table, nil
-}
-
-func (c *Cmd) getQbitState(instance int, app *apps.QbitConfig) (*State, error) { //nolint:cyclop
-	start := time.Now()
-	xfers, err := app.GetXfers()
-
-	state := &State{
-		Elapsed:  cnfg.Duration{Duration: time.Since(start)},
-		Instance: instance,
-		Name:     app.Name,
-		Next:     []*Sortable{},
-		Latest:   []*Sortable{},
-	}
-
-	if err != nil {
-		return state, fmt.Errorf("getting transfers from instance %d: %w", instance, err)
-	}
-
-	for _, xfer := range xfers {
-		if xfer.Eta != 8640000 && xfer.Eta != 0 && xfer.AmountLeft > 0 {
-			state.Next = append(state.Next, &Sortable{
-				Name: xfer.Name,
-				Date: time.Now().Add(time.Second * time.Duration(xfer.Eta)),
-			})
-		} else if xfer.AmountLeft == 0 {
-			state.Latest = append(state.Latest, &Sortable{
-				Name: xfer.Name,
-				Date: time.Unix(int64(xfer.CompletionOn), 0).Round(time.Second),
-			})
-		}
-
-		state.Size += xfer.Size
-		state.Uploaded += xfer.Uploaded
-		state.Downloaded += int64(xfer.Downloaded)
-		state.Downloads++
-
-		switch strings.ToLower(strings.TrimSpace(xfer.State)) {
-		case "stalledup", "moving", "forcedup":
-			state.Seeding++
-		case "downloading", "forceddl":
-			state.Downloading++
-		case "uploading":
-			state.Uploading++
-		case "pausedup", "pauseddl":
-			state.Paused++
-		case "queuedup", "checkingup", "allocating", "metadl", "queueddl", "stalleddl", "checkingdl", "checkingresumedata":
-			state.Incomplete++
-		case "unknown", "missingfiles", "error":
-			state.Errors++
-		default:
-			state.Errors++
-		}
-	}
-
-	sort.Sort(dateSorter(state.Next))
-	sort.Sort(sort.Reverse(dateSorter(state.Latest)))
-	state.Next.Shrink(showNext)
-	state.Latest.Shrink(showLatest)
-
-	return state, nil
-}
-
-func (c *Cmd) getRadarrState(instance int, r *apps.RadarrConfig) (*State, error) {
-	state := &State{Instance: instance, Next: []*Sortable{}, Latest: []*Sortable{}, Name: r.Name}
-	start := time.Now()
-
-	movies, err := r.GetMovie(0)
-	state.Elapsed.Duration = time.Since(start)
-
-	if err != nil {
-		return state, fmt.Errorf("getting movies from instance %d: %w", instance, err)
-	}
-
-	processRadarrState(state, movies)
-	sort.Sort(sort.Reverse(dateSorter(state.Latest)))
-	sort.Sort(dateSorter(state.Next))
-	state.Latest.Shrink(showLatest)
-	state.Next.Shrink(showNext)
-
-	return state, nil
-}
-
-func processRadarrState(state *State, movies []*radarr.Movie) { //nolint:cyclop
-	for _, movie := range movies {
-		state.Movies++
-		state.Size += movie.SizeOnDisk
-
-		if !movie.HasFile && movie.IsAvailable {
-			state.Missing++
-		}
-
-		if !movie.HasFile && !movie.IsAvailable {
-			state.Upcoming++
-		}
-
-		date := movie.DigitalRelease
-		if date.IsZero() || movie.PhysicalRelease.After(time.Now()) {
-			date = movie.PhysicalRelease
-		}
-
-		if date.After(time.Now()) && !movie.HasFile {
-			state.Next = append(state.Next, &Sortable{Name: movie.Title, Date: date})
-		}
-
-		if movie.MovieFile != nil {
-			state.Latest = append(state.Latest, &Sortable{Name: movie.Title, Date: movie.MovieFile.DateAdded})
-			state.OnDisk++
-		}
-	}
-}
-
-func (c *Cmd) getReadarrState(instance int, app *apps.ReadarrConfig) (*State, error) {
-	state := &State{Instance: instance, Next: []*Sortable{}, Name: app.Name}
-	start := time.Now()
-
-	books, err := app.GetBook("") // all books
-	state.Elapsed.Duration = time.Since(start)
-
-	if err != nil {
-		return state, fmt.Errorf("getting books from instance %d: %w", instance, err)
-	}
-
-	authorIDs := make(map[int64]struct{})
-
-	for _, book := range books {
-		have := false
-		state.Books++
-
-		if book.Statistics != nil {
-			authorIDs[book.AuthorID] = struct{}{}
-			state.Percent += book.Statistics.PercentOfBooks
-			state.Size += int64(book.Statistics.SizeOnDisk)
-			state.Editions += book.Statistics.TotalBookCount
-			state.Missing += int64(book.Statistics.BookCount - book.Statistics.BookFileCount)
-			have = book.Statistics.BookCount-book.Statistics.BookFileCount < 1
-			state.OnDisk += int64(book.Statistics.BookFileCount)
-		}
-
-		if book.ReleaseDate.After(time.Now()) && book.Monitored && !have {
-			state.Next = append(state.Next, &Sortable{
-				id:   book.ID,
-				Name: book.Title,
-				Date: book.ReleaseDate,
-				Sub:  book.AuthorTitle,
-			})
-		}
-	}
-
-	if state.Editions > 0 {
-		state.Percent /= float64(state.Editions)
-	} else {
-		state.Percent = 100
-	}
-
-	state.Authors = len(authorIDs)
-	sort.Sort(dateSorter(state.Next))
-	state.Next.Shrink(showNext)
-
-	if state.Latest, err = c.getReadarrHistory(app); err != nil {
-		return state, fmt.Errorf("instance %d: %w", instance, err)
-	}
-
-	return state, nil
-}
-
-// getReadarrHistory is not done.
-func (c *Cmd) getReadarrHistory(app *apps.ReadarrConfig) ([]*Sortable, error) {
-	history, err := app.GetHistoryPage(&starr.Req{
-		Page:     1,
-		PageSize: showLatest,
-		SortDir:  starr.SortDescend,
-		SortKey:  "date",
-		Filter:   readarr.FilterBookFileImported,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("getting history: %w", err)
-	}
-
-	table := []*Sortable{}
-
-	for idx := 0; idx < len(history.Records) && len(table) < showLatest; idx++ {
-		// An error here gets swallowed.
-		if book, err := app.GetBookByID(history.Records[idx].BookID); err == nil {
-			table = append(table, &Sortable{
-				Name: book.Title,
-				Sub:  book.AuthorTitle,
-				Date: history.Records[idx].Date,
-			})
-		}
-	}
-
-	return table, nil
-}
-
-func (c *Cmd) getSonarrState(instance int, app *apps.SonarrConfig) (*State, error) {
-	state := &State{Instance: instance, Next: []*Sortable{}, Name: app.Name}
-	start := time.Now()
-
-	allshows, err := app.GetAllSeries()
-	state.Elapsed.Duration = time.Since(start)
-
-	if err != nil {
-		return state, fmt.Errorf("getting series from instance %d: %w", instance, err)
-	}
-
-	for _, show := range allshows {
-		state.Shows++
-		if show.Statistics != nil {
-			state.Percent += show.Statistics.PercentOfEpisodes
-			state.Size += show.Statistics.SizeOnDisk
-			state.Episodes += int64(show.Statistics.TotalEpisodeCount)
-			state.Missing += int64(show.Statistics.EpisodeCount - show.Statistics.EpisodeFileCount)
-			state.OnDisk += int64(show.Statistics.EpisodeFileCount)
-		}
-
-		if show.NextAiring.After(time.Now()) {
-			state.Next = append(state.Next, &Sortable{
-				id:   show.ID,
-				Name: show.Title,
-				Date: show.NextAiring,
-			})
-		}
-	}
-
-	if state.Shows > 0 {
-		state.Percent /= float64(state.Shows)
-	} else {
-		state.Percent = 100
-	}
-
-	if state.Next, err = c.getSonarrStateUpcoming(app, state.Next); err != nil {
-		return state, fmt.Errorf("instance %d: %w", instance, err)
-	}
-
-	if state.Latest, err = c.getSonarrHistory(app); err != nil {
-		return state, fmt.Errorf("instance %d: %w", instance, err)
-	}
-
-	return state, nil
-}
-
-func (c *Cmd) getSonarrHistory(app *apps.SonarrConfig) ([]*Sortable, error) {
-	history, err := app.GetHistoryPage(&starr.Req{
-		Page:     1,
-		PageSize: showLatest + 5, //nolint:gomnd // grab extra in case there's an error.
-		SortDir:  starr.SortDescend,
-		SortKey:  "date",
-		Filter:   sonarr.FilterDownloadFolderImported,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("getting history: %w", err)
-	}
-
-	table := []*Sortable{}
-
-	for _, rec := range history.Records {
-		if len(table) >= showLatest {
-			break
-		}
-
-		series, err := app.GetSeriesByID(rec.SeriesID)
-		if err != nil {
-			continue
-		}
-
-		// An error here gets swallowed.
-		if eps, err := app.GetSeriesEpisodes(rec.SeriesID); err == nil {
-			for _, episode := range eps {
-				if episode.ID == rec.EpisodeID {
-					table = append(table, &Sortable{
-						Name:    series.Title,
-						Sub:     episode.Title,
-						Date:    rec.Date,
-						Season:  episode.SeasonNumber,
-						Episode: episode.EpisodeNumber,
-					})
-				}
-			}
-		}
-	}
-
-	return table, nil
-}
-
-func (c *Cmd) getSonarrStateUpcoming(app *apps.SonarrConfig, next []*Sortable) ([]*Sortable, error) {
-	sort.Sort(dateSorter(next))
-
-	redo := []*Sortable{}
-
-	for _, item := range next {
-		eps, err := app.GetSeriesEpisodes(item.id)
-		if err != nil {
-			return nil, fmt.Errorf("getting series ID %d (%s): %w", item.id, item.Name, err)
-		}
-
-		for _, episode := range eps {
-			if episode.AirDateUtc.Year() == item.Date.Year() &&
-				episode.AirDateUtc.YearDay() == item.Date.YearDay() &&
-				episode.SeasonNumber != 0 && episode.EpisodeNumber != 0 {
-				redo = append(redo, &Sortable{
-					Name:    item.Name,
-					Sub:     episode.Title,
-					Date:    episode.AirDateUtc,
-					Season:  episode.SeasonNumber,
-					Episode: episode.EpisodeNumber,
-				})
-
-				break
-			}
-		}
-
-		if len(redo) >= showNext {
-			break
-		}
-	}
-
-	return redo, nil
-}
-
-func (c *Cmd) getSabNZBStates() []*State {
-	states := []*State{}
-
-	for instance, app := range c.Apps.SabNZB {
-		if app.Timeout.Duration < 0 || app.URL == "" {
-			continue
-		}
-
-		c.Debugf("Getting SabNZB State: %d:%s", instance+1, app.URL)
-
-		state, err := c.getSabNZBState(instance+1, app)
-		if err != nil {
-			state.Error = err.Error()
-			c.Errorf("Getting SabNZB Data from %d:%s: %v", instance+1, app.URL, err)
-		}
-
-		states = append(states, state)
-	}
-
-	return states
-}
-
-func (c *Cmd) getSabNZBState(instance int, s *apps.SabNZBConfig) (*State, error) {
-	state := &State{Instance: instance, Name: s.Name}
-	start := time.Now()
-	queue, err := s.GetQueue(context.Background())
-	hist, err2 := s.GetHistory(context.Background())
-	state.Elapsed.Duration = time.Since(start)
-
-	if err != nil {
-		return state, fmt.Errorf("getting queue from instance %d: %w", instance, err)
-	} else if err2 != nil {
-		return state, fmt.Errorf("getting history from instance %d: %w", instance, err2)
-	}
-
-	state.Size = hist.TotalSize.Bytes
-	state.Month = hist.MonthSize.Bytes
-	state.Week = hist.WeekSize.Bytes
-
-	state.Downloads = len(queue.Slots) + hist.Noofslots
-	state.Next = []*Sortable{}
-	state.Latest = []*Sortable{}
-
-	for _, xfer := range queue.Slots {
-		if strings.EqualFold(xfer.Status, "Downloading") {
-			state.Downloading++
-		} else if strings.EqualFold(xfer.Status, "Paused") {
-			state.Paused++
-		}
-
-		if xfer.Mbleft > 0 {
-			state.Incomplete++
-		}
-
-		state.Next = append(state.Next, &Sortable{
-			Date: xfer.Eta.Round(time.Second).UTC(),
-			Name: xfer.Filename,
-		})
-	}
-
-	for _, xfer := range hist.Slots {
-		state.Latest = append(state.Latest, &Sortable{
-			Name: xfer.Name,
-			Date: time.Unix(xfer.Completed, 0).Round(time.Second).UTC(),
-		})
-
-		if xfer.FailMessage != "" {
-			state.Errors++
-		} else {
-			state.Downloaded++
-		}
-	}
-
-	sort.Sort(dateSorter(state.Next))
-	sort.Sort(sort.Reverse(dateSorter(state.Latest)))
-	state.Next.Shrink(showNext)
-	state.Latest.Shrink(showLatest)
-
-	return state, nil
 }
 
 type dateSorter []*Sortable
